use explicit_node
