+++ anvil/core.py
@@ -23,124 +23,6 @@ class TrainingRuncardNotFound(InvalidTrainingOutputError):
     pass
 
 
-class SpinHamiltonian(nn.Module):
-    """
-    Extend the nn.Module class to return the Hamiltonian for the classical
-    N-spin model (also known as the N-vector model), given either
-    a single state size (1, (N-1) * lattice_volume) or a stack of shape
-    (n_sample, (N-1) * lattice_volume).
-
-    The spins are defined as having modulus 1, such that they take values
-    on the (N-1)-sphere, and can be parameterised by N-1 angles using
-    spherical polar coordinates (with the radial coordinate equal to one).
-
-    Parameters
-    ----------
-    geometry:
-        define the geometry of the lattice, including dimension, size and
-        how the state is split into two parts
-    field_dimension: int
-        number of polar coordinates (angles) parameterising each spin vector
-        on the unit sphere. Note that this is equal to N-1 where N is the
-        number of Euclidean spin components!
-    beta: float
-        the inverse temperature (coupling strength).
-
-    Notes
-    -----
-    There are separate methods for the N = 2 (XY) and N = 3 (Heisenberg)
-    models, since the Hamiltonians can be written directly in terms of the
-    polar coordinates in a simple form.
-
-    For higher dimensional spin models, the spin vector components are first
-    computed from the polar coordinates.
-    """
-
-    def __init__(self, field_dimension, beta, geometry):
-        super().__init__()
-        self.field_dimension = field_dimension  # N-1
-        self.beta = beta
-        self.geometry = geometry
-        self.volume = self.geometry.length ** 2
-        self.shift = self.geometry.get_shift()
-
-        if self.field_dimension == 1:
-            self.forward = self.xy_hamiltonian
-        elif self.field_dimension == 2:
-            self.forward = self.heisenberg_hamiltonian
-        else:
-            self.forward = self.n_spin_hamiltonian
-
-    def xy_hamiltonian(self, state: torch.Tensor) -> torch.Tensor:
-        """
-        Compute XY Hamiltonian from a stack of angles (not Euclidean field components)
-        with shape (n_sample, lattice_volume).
-        """
-        hamiltonian = -self.beta * torch.cos(
-            state[:, self.shift] - state.view(-1, 1, self.volume)
-        ).sum(
-            dim=1,
-        ).sum(  # sum over two shift directions (+ve nearest neighbours)
-            dim=1, keepdim=True
-        )  # sum over lattice sites
-        return hamiltonian
-
-    def heisenberg_hamiltonian(self, state: torch.Tensor) -> torch.Tensor:
-        """
-        Compute classical Heisenberg Hamiltonian from a stack of angles with shape (n_sample, 2 * volume).
-
-        Reshapes state into shape (n_sample, lattice_volume, 2), so must make sure to keep this
-        consistent with observables.
-        """
-        polar = state[:, ::2]
-        azimuth = state[:, 1::2]
-        cos_polar = torch.cos(polar)
-        sin_polar = torch.sin(polar)
-
-        hamiltonian = -self.beta * (
-            cos_polar[:, self.shift] * cos_polar.view(-1, 1, self.volume)
-            + sin_polar[:, self.shift]
-            * sin_polar.view(-1, 1, self.volume)
-            * torch.cos(azimuth[:, self.shift] - azimuth.view(-1, 1, self.volume))
-        ).sum(
-            dim=1,
-        ).sum(  # sum over two shift directions (+ve nearest neighbours)
-            dim=1, keepdim=True
-        )  # sum over lattice sites
-        return hamiltonian
-
-    def _spher_to_eucl(self, state):
-        """
-        Take a stack of angles with shape (n_sample, (N-1) * lattice_volume), where the N-1
-        angles parameterise an N-spin vector on the unit (N-1)-sphere, and convert this
-        to a stack of euclidean field vectors with shape (n_sample, lattice_volume, N).
-        """
-        coords = state.view(-1, self.volume, self.field_dimension)
-
-        vector = torch.empty(coords.shape[0], self.volume, self.field_dimension + 1)
-        vector[:, :, :-1] = torch.cos(coords)
-        vector[:, :, 1:] *= torch.cumprod(torch.sin(coords), dim=2)
-
-        return vector
-
-    def n_spin_hamiltonian(self, state):
-        """
-        Compute the N-spin Hamiltonian from a stack of angles with shape (n_sample, field_dimension * volume).
-
-        """
-        field_vector = self._spher_to_eucl(state)
-
-        hamiltonian = -self.beta * torch.sum(
-            field_vector[:, self.shift, :]
-            * field_vector.view(-1, 1, self.volume, self.field_dimension + 1),
-            dim=-1,  # sum over vector components
-        ).sum(
-            dim=1,
-        ).sum(  # sum over shift directions
-            dim=1, keepdim=True
-        )  # sum over lattice sites
-        return hamiltonian
-
 
 class Checkpoint:
     """Class which saves and loads checkpoints and allows checkpoints to be
