+++ anvil/sample.py
@@ -65,6 +65,9 @@ def sample_batch(
         z, base_log_density = base(batch_size + 1)
         phi, map_log_density = loaded_model(z)  # map using trained loaded_model to phi
         model_log_density = base_log_density + map_log_density
+
+        np.savetxt("base.txt", z)
+        np.savetxt("target.txt", phi)
         
         if current_state is not None:
             phi[0] = current_state
@@ -173,9 +176,9 @@ def chain_autocorrelation(
     )
 
     accepted = float(torch.sum(history))
-    n_states = len(history)
+    sample_size = len(history)
     autocorrelations = torch.zeros(
-        n_states + 1, dtype=torch.float
+        sample_size + 1, dtype=torch.float
     )  # +1 in case 100% rejected
     consecutive_rejections = 0
 
@@ -195,7 +198,7 @@ def chain_autocorrelation(
 
     # Compute integrated autocorrelation
     integrated_autocorrelation = 0.5 + torch.sum(
-        autocorrelations / torch.arange(n_states + 1, 0, -1, dtype=torch.float)
+        autocorrelations / torch.arange(sample_size + 1, 0, -1, dtype=torch.float)
     )
     sample_interval = ceil(2 * integrated_autocorrelation)
     log.info(
